requirements.txt
pytest
openai
python-dotenv
pytest-xdist